# Optimize rgb565 serialization