--- a/requirements.txt
+++ b/requirements.txt
@@ -1,6 +1,7 @@
 # Python packages requirements
 Pillow~=9.5.0         # Image generation
 pyserial~=3.5         # Serial linl to communicate with the display
+numpy~=1.19           # Efficient image serialization
 PyYAML~=6.0           # For themes files
 psutil~=5.9.5         # CPU / disk / network metrics
 GPUtil~=1.4.0         # Nvidia GPU