For context: I'm trying to drive a rev A display from a (gen 1) Raspberry Pi. On a weak CPU like that, the sample program `simple-program.py` is extremely slow (~30 s to paint the background image) and the CPU usage is very high (constantly ~80 %).

So, I did some profiling and I've seen that the serialization to little-endian RGB565 was the main culprit for CPU usage. Using numpy yields a huge improvement.

I've added some timing debug messages to the `simple-program`, here's before:

```
2023-09-05 23:28:01.624 [DEBUG] background picture set (took 29.658 s)
2023-09-05 23:28:09.804 [DEBUG] refresh done (took 4.121 s)
2023-09-05 23:28:13.989 [DEBUG] refresh done (took 4.178 s)
2023-09-05 23:28:18.138 [DEBUG] refresh done (took 4.142 s)
2023-09-05 23:28:22.278 [DEBUG] refresh done (took 4.134 s)
```

and after:

```
2023-09-05 23:29:01.502 [DEBUG] background picture set (took 6.218 s)
2023-09-05 23:29:03.760 [DEBUG] refresh done (took 1.265 s)
2023-09-05 23:29:04.970 [DEBUG] refresh done (took 1.204 s)
2023-09-05 23:29:06.182 [DEBUG] refresh done (took 1.205 s)
2023-09-05 23:29:07.392 [DEBUG] refresh done (took 1.204 s)
```

Also, the CPU usage goes from ~80% to ~40%.

So, it's still not fast, but a 5x speedup in showing the BG, and >3x speedup in refreshing the info makes it much more useable.

On my desktop with a more powerful CPU, the timings only show a slight improvement (BG paint goes from 1.49s to 1.33s, refresh from 0.22s to 0.19s). However, there's a 2x difference in CPU usage, from ~30% to ~15%.

Sure, there's the downside of adding numpy as a dep, but I think that's not a big problem, lots of things depend on it, the Raspberry Pi OS even comes with it pre-installed.