@braewoods 2023-09-06
If CPU usage is this much of a concern, then the screen I am working on writing alternative software for may be a better option. All the drawing is done on the device end. See the main README.md. It is under the section called Fuldho.