@mathoudebine 2023-09-06
Thanks for optimizing this code! I agree adding numpy is not a concern as it is a well known and maintained package