@hchargois 2023-09-06
> Thanks for optimizing this code! I agree adding numpy is not a concern as it is a well known and maintained package

I only have a rev A screen, so I couldn't test the other, but I see that at least rev B also uses RGB565, so the same kind of optimization could be done. Rotating/reversing should also be easily done with numpy transposing/rolling operations, I think (I'm not a numpy expert at all).

I did some further testing and I noticed that my first post is totally misrepresenting the performance gains, which are much bigger than I thought!

In the current code, the serialization to RGB565 and the sending of data to the serial port are interlaced. To better measure the performance gains of the serialization only, I moved the current pure-python serialization in a first step, and the line sending in a second step:

```python
        pix = image.load()
        line = bytes()
        lines = []
        for h in range(image_height):
            for w in range(image_width):
                R = pix[w, h][0] >> 3
                G = pix[w, h][1] >> 2
                B = pix[w, h][2] >> 3

                rgb = (R << 11) | (G << 5) | B
                line += struct.pack('<H', rgb)

                # Send image data by multiple of "display width" bytes
                if len(line) >= self.get_width() * 8:
                    lines.append(line)
                    line = bytes()
        # Write last line if needed
        if len(line) > 0:
            lines.append(line)

        end = time.perf_counter()
        logger.debug(f"serialization done (took {end-start:.3f} s)")

        # Lock queue mutex then queue all the requests for the image data
        start = time.perf_counter()

        with self.update_queue_mutex:
            for line in lines:
                self.SendLine(line)

        end = time.perf_counter()
        logger.debug(f"sending lines done (took {end-start:.3f} s)")
```

On my Raspberry Pi, here are the results of the pure-python serialization for a fullscreen image:

```
serialization done (took 21.458 s)
sending lines done (took 5.921 s)
```

And here's the numpy version:

```
serialization done (took 0.160 s)
sending lines done (took 5.957 s)
```

So, it's **more than 100x faster**! It seems that numpy is crazy efficient for this kind of things.

Now, the paint speed is basically limited by the speed of writing the data to serial, which I don't know why it's so slow on the raspberry pi...