@braewoods 2023-09-06
> Now, the paint speed is basically limited by the speed of writing the data to serial, which I don't know why it's so slow on the raspberry pi...

You have to keep in mind most of the screens supported by this project are using a very slow USB 1.1 interface. A whole screen blit will take 2-3 seconds. But there's also a possibility that python is still acting as a noticeable bottleneck on your slower ARM board. Python is *very* slow after all, especially at raw number crunching as you are seeing here.

Perhaps in time there will be a better solution available. I've been thinking of starting another project to handle USB screens like the Turing ones at some point once my current `hidss` project is finished. The focus of that project would be to improve the performance of the driver program by designing a new driver program with a focus on performance optimization. This won't make much difference for PCs but it may make the screens usable by lower performing hardware, which would be my primary motivation.