@hchargois 2023-09-06
> You have to keep in mind most of the screens supported by this project are using a very slow USB 1.1 interface. A whole screen blit will take 2-3 seconds. 

Sure, USB 1.1 is slow, but it's still 12 Mbit/s, and the screen is recognized at that speed in both my Raspberry Pi and desktop PC.

Ignoring some USB overhead, at a full 12 Mbit/s, sending the 320*480*16 bits for a fullscreen paint could take on the order of 0.03 s, which is much lower than the numbers I'm seeing, so we're not really limited by the USB bandwidth itself here.

> But there's also a possibility that python is still acting as a noticeable bottleneck on your slower ARM board. Python is _very_ slow after all, especially at raw number crunching as you are seeing here.

After the serialization to RGB565, there's not really any "number crunching", it's just writing bytes to a file descriptor, which shouldn't really be impacted by Python's speed (or rather slowness).

But I did think the same thing as you! So to test that I implemented a minimal Go program to paint a background, and I got the exact same results as with Python: ~1.2s on my desktop PC, ~6s on the RPi. Some further stracing show that indeed the individual write system calls themselves are slower on the RPi than on the PC.

> Perhaps in time there will be a better solution available. I've been thinking of starting another project to handle USB screens like the Turing ones at some point once my current `hidss` project is finished. The focus of that project would be to improve the performance of the driver program by designing a new driver program with a focus on performance optimization. This won't make much difference for PCs but it may make the screens usable by lower performing hardware, which would be my primary motivation.

Yep, a "driver" backend in a fast language would be nice, I thought on doing the same.