@braewoods 2023-09-06
> Yep, a "driver" backend in a fast language would be nice, I thought on doing the same.

I would be using C most likely if I end up doing it. I also wanted to implement a "separation of concerns" of sorts so the whole program doesn't have to run as `root`. For `hidss` I am doing this by making a minimal controller program that will be `setuid` `root` and the rest of the programs will be unprivileged. Unfortunately you can't do this with a `python` program because most kernels require a `setuid` program to be a native executable and not a script that requires an interpreter.