@hchargois 2023-09-07
OK so this message has nothing to do with `turing-smart-screen-python` per se, but I'm happy to say that I've managed to make my screen work well with my Raspberry Pi. In the chance that someone else stumbles upon this thread trying to do the same thing, I'll share how I made it work.

As per the info in the Raspberry Pi forums [1] I simply added `dwc_otg.speed=1` to the `/boot/cmdline.txt` to set the USB host controller to USB 1.1 speed. My RPi is a 1 B, but AFAIK the same controller is used in RPIs 2 and 3 so I expect they will exhibit the same problem & solution.

With that parameter, I have fullscreen paints in around 1.5 s, so it's just marginally slower than my desktop PC (1.2 s).

Of course the Ethernet is now very slow by today's standards (~7 Mbit/s), but that's more than enough for my use case.

[1] https://forums.raspberrypi.com/viewtopic.php?t=53832