@mathoudebine 2023-09-14
Thanks @hchargois I added this info to the Troubleshooting page https://github.com/mathoudebine/turing-smart-screen-python/wiki/Troubleshooting#raspberry-pi-zero--1--2--3-display-refresh-is-too-slow